feat: Remove object_tag_class from Taxonomy
